Opening Hour Controller and edited Opening Hour Service

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/library/service/OpeningHourService.java
@@ -20,6 +20,28 @@ public class OpeningHourService {
 	@Transactional
 	public OpeningHour createOpeningHour(DayOfWeek day, Time startTime, Time endTime) {
 		
+		String error = "";
+		
+		if (day == null || day.equals("")) {
+			error = error + "Opening Hour day cannot be empty!";
+		}
+		if (startTime == null) {
+			error = error + "Opening Hour start time cannot be empty!";
+		}
+		
+		if (endTime == null) {
+			error = error + "Opening Hour end time cannot be empty!";
+		}
+		
+		if (endTime != null && startTime != null && endTime.before(startTime)) {
+	        error = error + "Opening Hour end time cannot be before event start time!";
+	    }
+		
+		error = error.trim();
+		if (error.length() > 0) {
+			throw new IllegalArgumentException(error);
+		}
+		
 		OpeningHour o = new OpeningHour();
 		o.setDay(day);
 		o.setStartTime(startTime);
@@ -42,6 +64,10 @@ public class OpeningHourService {
 	
 	@Transactional 
 	public OpeningHour deleteOpeningHour(OpeningHour o) {
+		if(o == null) {
+			throw new IllegalArgumentException("Inputed Opening Hour must not be null!");
+		}
+		
 		openingHourRepository.delete(o);
 		o = null;
 		return o;
@@ -49,6 +75,9 @@ public class OpeningHourService {
 	
 	@Transactional
 	public OpeningHour updateOpeningHourDayOfWeek(OpeningHour o, DayOfWeek day) {
+		if (day == null || day.equals("")) {
+			throw new IllegalArgumentException("Opening Hour day cannot be empty!");
+		}
 		openingHourRepository.delete(o);
 		o.setDay(day);
 		openingHourRepository.save(o);
@@ -57,6 +86,9 @@ public class OpeningHourService {
 	
 	@Transactional
 	public OpeningHour updateOpeningHourStartTime(OpeningHour o, Time startTime) {
+		if (startTime == null) {
+			throw new IllegalArgumentException("Opening Hour start time cannot be empty!");
+		}
 		openingHourRepository.delete(o);
 		o.setStartTime(startTime);
 		openingHourRepository.save(o);
@@ -65,6 +97,10 @@ public class OpeningHourService {
 	
 	@Transactional
 	public OpeningHour updateOpeningHourEndTime(OpeningHour o, Time endTime) {
+		if (endTime == null) {
+			throw new IllegalArgumentException("Opening Hour end time cannot be empty!");
+		}
+		
 		openingHourRepository.delete(o);
 		o.setEndTime(endTime);
 		openingHourRepository.save(o);
